move diagutils to internal
